Move the animacy section after the overview table
In the current version, the section explaining animacy is before the pronoun overview table. It makes more sense to have the overview table in the beginning of the section. The animacy section (which explains that there are animate and inanimate 3rd pronouns) fits in much better with the existing section on inclusivity (which explains that there are inclusive and exclusive 1st person pronouns).