Update about new used with subclauses
According to an example provided by KP and posted on the Kelutral server on September 15th, <iv> is not obligatory. Maybe it allows the distinction between an uncertain/wished action and an action that is actually done after the subject expresses a desire to do that. I prefer to word it like this since we don't have more information about it.